MemoryLayouts.jl 🧠⚡

Optimize your memory layout for maximum cache efficiency.

Documentation for MemoryLayouts.

🚀 The Problem vs. The Solution

Standard collections in Julia (Dicts, Arrays of Arrays, structs) often scatter data across memory, causing frequent cache misses. MemoryLayouts.jl packs this data into contiguous blocks.

🔮 How it works

Function	Description	Analogy

layout( x )	Aligns immediate fields of x	Like copy( x ) but packed

deeplayout( x )	Recursively aligns nested structures	Like deepcopy( x ) but packed

layoutstats( x )	Dry run statistics for layout( x )

deeplayoutstats( x )	Dry run statistics for deeplayout( x )

🛠️ Usage

The package provides four exported functions: layout, deeplayout, layoutstats and deeplayoutstats. The distinction between the first two functions is that layout only applies to top level objects, whereas deeplayout applies to objects at all levels. The two examples below demonstrate their use. As for the stats functions, these just do a dry run and print out some statistics on the degree of contiguity improvement a user can expect to see.

💡 Example for layout

The example below demonstrates how to use layout.

using MemoryLayouts, BenchmarkTools, StyledStrings

function original( A = 10_000, L = 100, S = 5000)
    x = Vector{Vector{Float64}}( undef, A )
    s = Vector{Vector{Float64}}( undef, A )
    for i ∈ 1:A
        x[i] = rand( L )
        s[i] = rand( S )
    end
    return x
end

function computeme( X )
    Σ = 0.0
    for x ∈ X 
        Σ += x[5] 
    end
    return Σ
end

print( styled"{(fg=0xff9999):original}: " ); @btime computeme( X ) setup=(X = original(););
print( styled"{(fg=0x99ff99):layout}: " ); @btime computeme( X ) setup=(X = layout( original()););
;

💡 Example for deeplayout

The example below illustrates the use of deeplayout.

using MemoryLayouts, BenchmarkTools, StyledStrings

struct 𝒮{X,Y,Z}
    x :: X
    y :: Y 
    z :: Z
end

function original( A = 10_000, L = 100, S = 5000)
    x = Vector{Vector{Float64}}( undef, A )
    s = Vector{Vector{Float64}}( undef, A )
    for i ∈ 1:A
        x[i] = rand( L )
        s[i] = rand( S )
    end
    return 𝒮( [x[i] for i ∈ 1:div( A, 3 )], [ x[i] for i ∈ div( A, 3 )+1:div( 2*A, 3 )], [x[i] for i ∈ div( 2*A, 3 )+1:A ] )
end

function computeme( X )
    Σ = 0.0
    for x ∈ X.x  
        Σ += x[5] 
    end
    for y ∈ X.y 
        Σ += y[37]
    end
    for z ∈ X.z 
        Σ += z[5] 
    end
    return Σ
end

println( layoutstats( original() ) )
println( deeplayoutstats( original() ) )

print( styled"{(fg=0xff9999):original}: " ); @btime computeme( X ) setup=(X = original(););
print( styled"{(fg=0x99ff99):layout}: " ); @btime computeme( X ) setup=(X = layout( original()););
print( styled"{(fg=0x9999ff):deeplayout}: " ); @btime computeme( X ) setup=(X = deeplayout( original()););
;

📊 Dry Run / Statistics

You can inspect the potential improvements in memory contiguity without performing the actual allocation using layoutstats and deeplayoutstats.

julia> using MemoryLayouts

julia> data = [rand(10) for _ in 1:5];

julia> layoutstats(data)
LayoutStats(packed=400 b, blocks=5, span=2 kb, reduction=2 kb (82.6%))
  Level 1: bytes=400 b, blocks=5, span=2 kb, reduction=2 kb (82.6%)

The output indicates:

packed: The total size (in bytes) of the data if packed.

blocks: The number of individual arrays identified.

span: The current distance between the minimum and maximum memory addresses of the data.

reduction: The potential reduction in memory span.

🤝 Compatibility

MemoryLayouts.jl is compatible with

AxisKeys

InlineStrings

NamedDims.jl

OffsetArrays

this assumes that those packages are loaded by the user (weak dependences)

⚡ SIMD Alignment

Both layout and deeplayout accept an optional alignment keyword argument (default 1). This allows you to specify the byte alignment for the start of each array in the contiguous memory block.

Proper memory alignment is relevant for maximizing performance with SIMD (Single Instruction, Multiple Data) instructions (e.g., AVX2, AVX-512). On the other hand, such alignment leaves gaps between blocks of memory that are not a multiple of 64 bytes in length.

AVX2 typically requires 32-byte alignment.

AVX-512 typically requires 64-byte alignment.

💡 Example

using MemoryLayouts
struct MyData
    a::Vector{Float64}
    b::Vector{Float64}
end

data = MyData( rand( 100 ), rand( 100 ) )

aligneddata = layout( data; alignment = 64 )

pointer( aligneddata.a ) # Will be a multiple of 64
pointer( aligneddata.b ) # Will be a multiple of 64

⚠️ Things to be mindful of

Important details

it operates on various types of collections including structs, arrays, and dicts

operating on means that these collections are traversed, possibly recursively

the only objects that are copied into contiguous memory are isbits arrays (think arrays of numbers, InlineStrings (but not regular strings), etcetera )

the more scattered is the memory before the layout change, the greater is the potential speed gain

layout copies, but only the top level; see example 2 above

deeplayout copies all levels

no attempt is made to make empty arrays contiguous

no attempt is made to make objects that are not one of the covered collections contiguous

the package assigns one memory block and within that block uses unsafe_wraps to obtain Julia arrays

this can have 'interesting' consequences if misused

ergo, this package should not be used by those new to programming

objects can be excluded from layout changes via the exclude keyword

there is overhead in laying out memory initially and (to a much lesser extent) to running the finalizer

thus, MemoryLayouts works best for aligning memory in a collection once and then using it for an extended stretch

resizing one of the arrays whose memory was laid out by MemoryLayouts is safe, but likely results in that array being moved to another location in memory assigned by Julia (not by MemoryLayouts)

reassigning an array assigned by MemoryLayouts to another location, e.g. by writing y[i] = ... does not release the entire memory block

the entire memory block is only released if the entire collection loses scope

by default, MemoryLayouts packs in the isbits arrays as tightly as it can

this may not be optimal, e.g. for AVX-512 computations

use the alignment=64 option to give up some contiguity and regain alignment desired for optimal AVX-512 performance

📖 Function documentation

MemoryLayouts.layout — Function

layout(s; exclude = Symbol[], alignment::Int=1)

layout aligns the memory of arrays within the object s, whose type should be one of struct, AbstractArray, or AbstractDict

layout creates a new instance of s (or copy of s) where the arrays are stored contiguously in memory.

The alignment keyword argument specifies the memory alignment in bytes. This is particularly useful for SIMD operations, where aligning data to 16, 32, or 64 bytes can improve performance.

Excluded items are preserved as-is (or deep-copied in some contexts) but not packed into the contiguous memory block.

important implementation details

Users should be mindful of the following important implementation details:

aligned arrays share a single contiguous memory block

resizing any of the arrays (push!, append!) will break this contiguity for that array (it will be reallocated elsewhere)

Contiguity is maintained until an array is resized or reassigned

MemoryLayouts.deeplayout — Function

deeplayout( x; exclude = Symbol[], alignment::Int=1 )

deeplayout recursively aligns memory of arrays within x and its fields

Unlike layout, which only aligns the immediate fields/elements of x, deeplayout traverses the structure recursively. In other words, deeplayout is to layout what deepcopy is to copy.

The alignment keyword argument specifies the memory alignment in bytes. This is particularly useful for SIMD operations, where aligning data to 16, 32, or 64 bytes can improve performance.

Excluded items are preserved as-is (or deep-copied in some contexts) but not packed into the contiguous memory block.

important implementation details

Users should be mindful of the following important implementation details:

aligned arrays share a single contiguous memory block

resizing any of the arrays (push!, append!) will break this contiguity for that array (it will be reallocated elsewhere)

Contiguity is maintained until an array is resized or reassigned

MemoryLayouts.layoutstats — Function

layoutstats( s; exclude = Symbol[], alignment :: Int = 1 )

Returns a LayoutStats object containing statistics about the memory layout if layout( s ) were called.

The returned object includes:

bytes: Total size (in bytes) of the data that would be packed.

blocks: Number of individual arrays identified.

span: The distance between the minimum and maximum memory addresses of the data.

reduction: The potential reduction in memory span (span - bytes).

MemoryLayouts.deeplayoutstats — Function

deeplayoutstats( x; exclude = Symbol[], alignment :: Int = 1 )

Returns a LayoutStats object containing statistics about the memory layout if deeplayout( x ) were called.

The returned object includes:

bytes: Total size (in bytes) of the data that would be packed.

blocks: Number of individual arrays identified.

span: The distance between the minimum and maximum memory addresses of the data.

reduction: The potential reduction in memory span (span - bytes).